Add files via upload
Contact Management System project is written in Python. The project file contains a python script. This is a simple GUI based project which is very easy to understand and use. Talking about the system, it contains all the required functions which include adding, viewing, deleting and updating contact lists. While adding the contact of a person, he/she has to provide first name, last name, gender, address and contact details. The user can also update the contact list if he/she wants to. For this, the user has to double-click on a record that he/she wishes to edit. The system shows the contact details in a list view. And also the user easily delete any contact details.

This GUI based Contact Management system provides the simplest management of contact details. In short, this projects mainly focus on CRUD. There’s an external database connection file used in this mini project to save user’s data permanently. In order to run the project, you must have installed Python, on your PC. This is a simple GUI Based system .